requirements.txt
pandas
chromadb
langchain
langchain_core
langchain-community
langchain-chroma
langchain-openai
langchain-classic
pypdf
openai
python-dotenv
tiktoken
faiss-cpu
rank_bm25